refactor(dh): remove mock libs
