--- libs/dh/charges/data-access-mocks/README.md
@@ -1,1 +1,0 @@
-# Charges mocks data access
