--- libs/dh/charges/data-access-mocks/project.json
@@ -1,26 +1,0 @@
-{
-  "projectType": "library",
-  "root": "libs/dh/charges/data-access-mocks",
-  "sourceRoot": "libs/dh/charges/data-access-mocks/src",
-  "prefix": "dh",
-  "targets": {
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/dh/charges/data-access-mocks"],
-      "options": {
-        "jestConfig": "libs/dh/charges/data-access-mocks/jest.config.js",
-        "passWithNoTests": true
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "options": {
-        "lintFilePatterns": [
-          "libs/dh/charges/data-access-mocks/src/**/*.ts",
-          "libs/dh/charges/data-access-mocks/src/**/*.html"
-        ]
-      }
-    }
-  },
-  "tags": ["product:dh", "domain:charges", "type:data-access"]
-}
