--- libs/dh/metering-point/data-access-mocks/README.md
@@ -1,1 +1,0 @@
-# Metering Point mocks data access
