+++ workspace.json
@@ -6,7 +6,6 @@
     "app-ett": "apps/ett/app-ett",
     "dh-auth-feature-msal": "libs/dh/auth/feature-msal",
     "dh-charges-data-access-api": "libs/dh/charges/data-access-api",
-    "dh-charges-data-access-mocks": "libs/dh/charges/data-access-mocks",
     "dh-core-shell": "libs/dh/core/shell",
     "dh-globalization-assets-localization": "libs/dh/globalization/assets-localization",
     "dh-globalization-configuration-localization": "libs/dh/globalization/configuration-localization",
@@ -14,7 +13,6 @@
     "dh-globalization-domain": "libs/dh/globalization/domain",
     "dh-globalization-feature-language-picker": "libs/dh/globalization/feature-language-picker",
     "dh-metering-point-data-access-api": "libs/dh/metering-point/data-access-api",
-    "dh-metering-point-data-access-mocks": "libs/dh/metering-point/data-access-mocks",
     "dh-metering-point-domain": "libs/dh/metering-point/domain",
     "dh-metering-point-feature-identity-and-master-data": "libs/dh/metering-point/feature-identity-and-master-data",
     "dh-metering-point-feature-overview": "libs/dh/metering-point/feature-overview",
